Move specs to module folders
