+++ src/js/modules/comments/sagas.spec.ts
@@ -1,0 +1,14 @@
+// tslint:disable:no-object-literal-type-assertion
+
+describe('Comments sagas', () => {
+  // const api = createApi();
+  // const sagaFunctions = createSagas(api).functions;
+
+  describe('createComment', () => {
+    it('TODO');
+  });
+
+  describe('deleteComment', () => {
+    it('TODO');
+  });
+});
