+++ src/js/modules/user/sagas.spec.ts
@@ -1,0 +1,26 @@
+// tslint:disable:no-object-literal-type-assertion
+
+describe('User sagas', () => {
+  // const api = createApi();
+  // const sagaFunctions = createSagas(api).functions;
+
+  describe('login', () => {
+    // TODO
+  });
+
+  describe('logout', () => {
+    // TODO
+  });
+
+  describe('redirectToLogin', () => {
+    // TODO
+  });
+
+  describe('signupUser', () => {
+    // TODO
+  });
+
+  describe('loadTeamInformation', () => {
+    // TODO
+  });
+});
